Rename p5iq-locate-function to p5iq-locate-sub
Since sub can mean both "function" or "method" in Perl5, it is also not
trivial to decide whether a 'sub' is really a 'method' when seeing the
definition of it. So there are going to be many false positive when
trying to distinguish those cases. Might as well just merge them.